fix: linting resolve

+++ src/test/AtomicLaunchTest.t.sol
@@ -45,9 +45,18 @@ interface ICurvePool {
 
     function balances(uint256 arg0) external view returns (uint256);
 
-    function exchange(int128 i, int128 j, uint256 _dx, uint256 _min_dy) external;
+    function exchange(
+        int128 i,
+        int128 j,
+        uint256 _dx,
+        uint256 _min_dy
+    ) external;
 
-    function get_dy(int128 i, int128 j, uint256 _dx) external returns (uint256);
+    function get_dy(
+        int128 i,
+        int128 j,
+        uint256 _dx
+    ) external returns (uint256);
 }
 
 contract AtomicLaunchTest is BaseFixture {
@@ -655,7 +664,7 @@ contract AtomicLaunchTest is BaseFixture {
         knightingRoundClaim(knightingRound_badger, influence_user);
         knightingRoundClaim(knightingRoundWithEth, eth_user);
 
-        // user wants to stake citadel 
+        // user wants to stake citadel
         vm.prank(address(citadelMinter));
         citadel.mint(btc_user, 100e18);
 
@@ -664,20 +673,23 @@ contract AtomicLaunchTest is BaseFixture {
         citadel.approve(address(xCitadel), 10e18);
         // deposit
         xCitadel.deposit(10e18);
-        uint userxCitadelBalance = xCitadel.balanceOf(btc_user);
+        uint256 userxCitadelBalance = xCitadel.balanceOf(btc_user);
         uint256 expectedClaimableBalance = (xCitadel.balance() *
             userxCitadelBalance) / xCitadel.totalSupply();
         xCitadel.withdrawAll();
         vm.warp(block.timestamp + xCitadelVester.INITIAL_VESTING_DURATION());
-        
-        uint userCitadelBefore = citadel.balanceOf(btc_user);
+
+        uint256 userCitadelBefore = citadel.balanceOf(btc_user);
         xCitadelVester.claim(btc_user, expectedClaimableBalance);
-        uint userCitadelAfter = citadel.balanceOf(btc_user);
+        uint256 userCitadelAfter = citadel.balanceOf(btc_user);
 
-        assertEq(userCitadelAfter-userCitadelBefore, expectedClaimableBalance);
+        assertEq(
+            userCitadelAfter - userCitadelBefore,
+            expectedClaimableBalance
+        );
 
-        uint citadelPoolBalanceBefore = pool.balances(0);
-        uint wbtcPoolBalanceBefore = pool.balances(1);
+        uint256 citadelPoolBalanceBefore = pool.balances(0);
+        uint256 wbtcPoolBalanceBefore = pool.balances(1);
 
         // provide liquidity to pool
 
@@ -687,15 +699,15 @@ contract AtomicLaunchTest is BaseFixture {
         citadel.approve(address(pool), 10e18);
         pool.add_liquidity(amounts, 0);
 
-        uint citadelPoolBalanceAfter = pool.balances(0);
-        uint wbtcPoolBalanceAfter = pool.balances(1);
+        uint256 citadelPoolBalanceAfter = pool.balances(0);
+        uint256 wbtcPoolBalanceAfter = pool.balances(1);
 
-        assertEq(citadelPoolBalanceAfter-citadelPoolBalanceBefore, 10e18);
-        assertEq(wbtcPoolBalanceAfter, wbtcPoolBalanceBefore) ;
+        assertEq(citadelPoolBalanceAfter - citadelPoolBalanceBefore, 10e18);
+        assertEq(wbtcPoolBalanceAfter, wbtcPoolBalanceBefore);
 
         // swap tokens
-        uint wbtcUserBalanceBefore = wbtc.balanceOf(btc_user);
-        uint userCitadelBalanceBefore = citadel.balanceOf(btc_user);
+        uint256 wbtcUserBalanceBefore = wbtc.balanceOf(btc_user);
+        uint256 userCitadelBalanceBefore = citadel.balanceOf(btc_user);
         citadelPoolBalanceBefore = pool.balances(0);
         wbtcPoolBalanceBefore = pool.balances(1);
         citadel.approve(address(pool), 20e18);
@@ -704,31 +716,38 @@ contract AtomicLaunchTest is BaseFixture {
         pool.exchange(0, 1, 20e18, 0);
         citadelPoolBalanceAfter = pool.balances(0);
         wbtcPoolBalanceAfter = pool.balances(1);
-        uint userCitadelBalanceAfter = citadel.balanceOf(btc_user);
+        uint256 userCitadelBalanceAfter = citadel.balanceOf(btc_user);
 
-        uint wbtcUserBalanceAfter = wbtc.balanceOf(btc_user);
-        emit log_named_uint("citadelPoolBalanceAfter" , citadelPoolBalanceAfter);
-        emit log_named_uint("citadelPoolBalanceBefore", citadelPoolBalanceBefore);
-        emit log_named_uint("userCitadelBalanceBefore" , userCitadelBalanceBefore);
+        uint256 wbtcUserBalanceAfter = wbtc.balanceOf(btc_user);
+        emit log_named_uint("citadelPoolBalanceAfter", citadelPoolBalanceAfter);
+        emit log_named_uint(
+            "citadelPoolBalanceBefore",
+            citadelPoolBalanceBefore
+        );
+        emit log_named_uint(
+            "userCitadelBalanceBefore",
+            userCitadelBalanceBefore
+        );
         emit log_named_uint("userCitadelBalanceAfter", userCitadelBalanceAfter);
-        emit log_named_uint("wbtcUserBalanceAfter" , wbtcUserBalanceAfter);
+        emit log_named_uint("wbtcUserBalanceAfter", wbtcUserBalanceAfter);
         emit log_named_uint("wbtcUserBalanceBefore", wbtcUserBalanceBefore);
-        emit log_named_uint("wbtcPoolBalanceBefore" , wbtcPoolBalanceBefore);
+        emit log_named_uint("wbtcPoolBalanceBefore", wbtcPoolBalanceBefore);
         emit log_named_uint("wbtcPoolBalanceAfter", wbtcPoolBalanceAfter);
         // assertEq(citadelPoolBalanceAfter-citadelPoolBalanceBefore, 20e18);
 
-        assertEq(wbtcUserBalanceAfter-wbtcUserBalanceBefore, wbtcPoolBalanceBefore-wbtcPoolBalanceAfter);
+        assertEq(
+            wbtcUserBalanceAfter - wbtcUserBalanceBefore,
+            wbtcPoolBalanceBefore - wbtcPoolBalanceAfter
+        );
     }
 
-    function knightingRoundClaim(
-        KnightingRound round,
-        address user) internal{
-            uint userBalanceBefore = xCitadel.balanceOf(user);
-            vm.prank(user);
-            uint tokenOutAmount = round.claim();
+    function knightingRoundClaim(KnightingRound round, address user) internal {
+        uint256 userBalanceBefore = xCitadel.balanceOf(user);
+        vm.prank(user);
+        uint256 tokenOutAmount = round.claim();
 
-            uint userBalanceAfter = xCitadel.balanceOf(user);
-            // check if user has received xCitadel
-            assertEq(tokenOutAmount, userBalanceAfter- userBalanceBefore);
+        uint256 userBalanceAfter = xCitadel.balanceOf(user);
+        // check if user has received xCitadel
+        assertEq(tokenOutAmount, userBalanceAfter - userBalanceBefore);
     }
 }
